Rename DockerAPI to Orchestrator
